fix(platform): show budget limit banner on both chat start and thread pages
Extract budget banner into standalone BudgetBanner component rendered
outside the scroll container so it stays fixed at the top. Banner is
now visible on thread pages (not just new chat) and always dismissible.
